tweet_collector/requirements.txt
pymongo
tweepy==4.4.0
dnspython
python-dotenv